refactored building reason functions
