+++ taipy/core/sequence/_sequence_manager.py
@@ -18,7 +18,6 @@ from .._entity._entity_ids import _EntityIds
 from .._manager._manager import _Manager
 from .._version._version_mixin import _VersionMixin
 from ..common._utils import _Subscriber
-from ..common.reason import Reason
 from ..common.warn_if_inputs_not_ready import _warn_if_inputs_not_ready
 from ..exceptions.exceptions import (
     InvalidSequence,
@@ -32,6 +31,8 @@ from ..job._job_manager_factory import _JobManagerFactory
 from ..job.job import Job
 from ..notification import Event, EventEntityType, EventOperation, Notifier
 from ..notification.event import _make_event
+from ..reason._reason_factory import _build_not_submittable_entity_reason
+from ..reason.reason import Reason
 from ..scenario._scenario_manager_factory import _ScenarioManagerFactory
 from ..scenario.scenario import Scenario
 from ..scenario.scenario_id import ScenarioId
@@ -350,7 +351,7 @@ class _SequenceManager(_Manager[Sequence], _VersionMixin):
         if not isinstance(sequence, Sequence):
             sequence = str(sequence)
             reason = Reason(sequence)
-            reason._add_reason(sequence, cls._build_not_submittable_entity_reason(sequence))
+            reason._add_reason(sequence, _build_not_submittable_entity_reason(sequence))
             return reason
 
         return sequence.is_ready_to_run()
