+++ taipy/core/task/_task_manager.py
@@ -20,13 +20,18 @@ from .._orchestrator._abstract_orchestrator import _AbstractOrchestrator
 from .._repository._abstract_repository import _AbstractRepository
 from .._version._version_manager_factory import _VersionManagerFactory
 from .._version._version_mixin import _VersionMixin
-from ..common.reason import Reason
 from ..common.warn_if_inputs_not_ready import _warn_if_inputs_not_ready
 from ..config.task_config import TaskConfig
 from ..cycle.cycle_id import CycleId
 from ..data._data_manager_factory import _DataManagerFactory
 from ..exceptions.exceptions import NonExistingTask
 from ..notification import EventEntityType, EventOperation, Notifier, _make_event
+from ..reason._reason_factory import (
+    _build_data_node_is_being_edited_reason,
+    _build_data_node_is_not_written,
+    _build_not_submittable_entity_reason,
+)
+from ..reason.reason import Reason
 from ..scenario.scenario_id import ScenarioId
 from ..sequence.sequence_id import SequenceId
 from ..submission.submission import Submission
@@ -170,16 +175,16 @@ class _TaskManager(_Manager[Task], _VersionMixin):
         if not isinstance(task, Task):
             task = str(task)
             reason = Reason(task)
-            reason._add_reason(task, cls._build_not_submittable_entity_reason(task))
+            reason._add_reason(task, _build_not_submittable_entity_reason(task))
         else:
             reason = Reason(task.id)
             data_manager = _DataManagerFactory._build_manager()
             for node in task.input.values():
                 node = data_manager._get(node)
                 if node._edit_in_progress:
-                    reason._add_reason(node.id, node._build_edit_in_progress_reason())
+                    reason._add_reason(node.id, _build_data_node_is_being_edited_reason(node.id))
                 if not node._last_edit_date:
-                    reason._add_reason(node.id, node._build_not_written_reason())
+                    reason._add_reason(node.id, _build_data_node_is_not_written(node.id))
 
         return reason
