+++ tests/core/common/test_reason.py
@@ -9,7 +9,7 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from taipy.core.common.reason import Reason
+from taipy.core.reason.reason import Reason
 
 
 def test_create_reason():
